+++ src/app/blog/[category]/[post]/_components/PostContainer.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import HighlightCode from "@/components/HighlightCode";
-import { PostMetaData } from "@/lib/types";
 import "@/styles/github.css";
 import "@/styles/highlight.css";
 import "@/styles/post.css";
+import type { PostMetaData } from "@/types/posts.types";
 import { useEffect, useState } from "react";
 
 interface HighlightedCodeProps {
