+++ src/app/blog/_components/PostSection/PostSectionMain/index.tsx
@@ -1,4 +1,4 @@
-import { Post } from "@/lib/types";
+import type { Post } from "@/types/posts.types";
 import { JSX, memo, Suspense, useCallback, useState } from "react";
 import { POSTSECTION_TEXT } from "..";
 import ArticleViewPost from "./ArticleViewPost";
