+++ src/scripts/buildCategory.ts
@@ -1,11 +1,10 @@
+import { listDirs, listFilesByExt, writeFileUtf8 } from "@/lib/fsUtils";
+import { CATEGORY_JSON_PATH, CONTENT_DIR } from "@/lib/paths";
+import type { CategoryList } from "@/types/posts.types";
 import path from "path";
-import { listFilesByExt, writeFileUtf8 } from "../fsUtils.js";
-import { CATEGORY_JSON_PATH, CONTENT_DIR } from "../paths.js";
-import { readCategories } from "../readers.js";
-import type { CategoryList } from "../types.js";
 
 const buildCategoryJson = () => {
-  const categories = readCategories(CONTENT_DIR);
+  const categories = listDirs(CONTENT_DIR);
   const result: CategoryList = categories.map((category) => {
     const categoryPath = path.join(CONTENT_DIR, category);
     const mdFiles = listFilesByExt(categoryPath, ".md");
