+++ src/types/posts.types.d.ts
